requirements.txt
absl-py==2.3.1
accelerate==1.12.0
aiocache==0.12.3
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.17.2
aniso8601==10.0.1
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.12.0
APScheduler==3.10.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.11.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.4.0
Authlib==1.6.5
av==16.0.1
azure-ai-documentintelligence==1.0.2
azure-core==1.37.0
azure-identity==1.25.0
azure-storage-blob==12.24.1
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
bidict==0.23.1
bitsandbytes==0.49.0
black==25.11.0
blinker==1.9.0
boto3==1.41.5
botocore==1.41.6
Bottleneck==1.6.0
brotli==1.2.0
build==1.3.0
cachetools==6.2.3
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
chromadb==1.0.20
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
compressed-tensors==0.13.0
contourpy==1.3.3
cryptography==46.0.3
ctranslate2==4.6.2
cycler==0.12.1
dataclasses-json==0.6.7
datasets==4.4.2
ddgs==9.9.2
decord==0.6.0
defusedxml==0.7.1
diffusers @ file:///D:/Carpeta%20de%20Juan/ModelosIA/newbie-image-exp/diffusers
dill==0.4.0
distro==1.9.0
docstring_parser==0.17.0
docx2txt==0.8
durationpy==0.10
ecdsa==0.19.1
einops==0.8.1
emoji==2.15.0
et_xmlfile==2.0.0
Events==0.5
fake-useragent==2.2.0
fastapi==0.123.0
faster-whisper==1.1.1
filelock==3.20.0
filetype==1.2.0
Flask==3.1.2
Flask-RESTful==0.3.10
flatbuffers==25.9.23
fonttools==4.61.1
fpdf2==2.8.2
frozenlist==1.8.0
fsspec==2025.10.0
ftfy==6.3.1
gallery_dl==1.31.1
gast==0.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
google-cloud-core==2.5.0
google-cloud-storage==2.19.0
google-crc32c==1.8.0rc4
google-genai==1.52.0
google-generativeai==0.8.5
google-pasta==0.2.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
h5py==3.15.1
hf-xet==1.2.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.34.0
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==3.13.0
keye-vl-utils==1.5.2
kiwisolver==1.4.9
kornia==0.8.2
kornia_rs==0.1.10
kubernetes==34.1.0
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.80
langchain-text-splitters==0.3.11
langdetect==1.0.9
langsmith==0.4.59
ldap3==2.9.1
libclang==18.1.1
loguru==0.7.3
longcat_image @ git+https://github.com/meituan-longcat/LongCat-Image.git@f756cc602d3268429ac4473df5998f24ee4d99e8
lxml==6.0.2
Mako==1.3.10
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.8
mcp==1.22.0
mdurl==0.1.2
ml_dtypes==0.5.4
mmh3==5.2.0
mpmath==1.3.0
msal==1.34.0
msal-extensions==1.3.1
msoffcrypto-tool==5.4.2
multidict==6.7.0
multiprocess==0.70.18
mypy_extensions==1.1.0
namex==0.1.0
networkx==3.6
nltk==3.9.1
numpy==2.2.6
oauthlib==3.3.1
olefile==0.47
ollama==0.6.1
onnx==1.20.0
onnxruntime==1.20.1
onnxruntime-gpu==1.23.2
open-webui==0.6.41
openai==2.11.0
opencv-python==4.12.0.88
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opensearch-py==2.8.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opt_einsum==3.4.0
optree==0.18.0
orjson==3.11.5
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
peewee==3.18.3
peewee-migrate==1.14.3
peft==0.18.0
pillow==11.3.0
platformdirs==4.5.1
posthog==5.4.0
primp==0.15.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pyclipper==1.4.0
pycparser==2.23
pycrdt==0.12.25
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
pymdown-extensions==10.17.2
PyMySQL==1.1.1
pypandoc==1.16.2
pyparsing==3.2.5
pypdf==6.4.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.12.3
python-iso639==2025.11.16
python-jose==3.5.0
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
python-socketio==5.15.0
pytokens==0.3.0
pytube==15.0.0
pytz==2025.2
pyvips==3.1.1
pywin32==311
pyxlsb==1.0.10
PyYAML==6.0.3
qwen-vl-utils==0.0.14
rank-bm25==0.2.2
RapidFuzz==3.14.3
rapidocr-onnxruntime==1.4.4
redis==7.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
RestrictedPython==8.0
rich==13.9.4
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.15.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==5.1.2
sentencepiece==0.2.1
shapely==2.1.2
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
soundfile==0.13.1
soupsieve==2.8
SQLAlchemy==2.0.38
sse-starlette==3.0.4
starlette==0.50.0
starlette-compress==1.6.1
starsessions==2.2.1
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
threadpoolctl==3.6.0
tiktoken==0.12.0
timm==1.0.22
tokenizers==0.22.1
torch==2.9.1+cu126
torchaudio==2.9.1+cu126
torchvision==0.24.1+cu126
tqdm==4.67.1
transformers==4.57.3
typer==0.20.0
typer-slim==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
unstructured==0.18.21
unstructured-client==0.42.4
uritemplate==4.2.0
urllib3==2.3.0
uuid_utils==0.12.0
uv==0.9.18
uvicorn==0.37.0
validators==0.35.0
watchfiles==1.1.1
wcwidth==0.2.14
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.4
win32_setctime==1.2.0
wrapt==2.0.1
wsproto==1.3.2
xlrd==2.0.1
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.22.0
youtube-transcript-api==1.2.2
yt-dlp==2025.12.8
yt-dlp-ejs==0.3.2
zipp==3.23.0
zstandard==0.25.0
